Create nice_number
Let's call a number nice if it is four-digit and divisible by 77 or 1717. Write a program to determine if an entered number is nice.
The program should print "YES" if the number is nice, or "NO" otherwise.

Input data format: A natural number is fed to the program.
Output data format: The program should output the text in accordance with the condition of the problem.