Create leap-year
A program that determines if a year with a given number is a leap year. If the year is a leap year, then output "YES", otherwise output "NO". A year is a leap year if its number is a multiple of 4 but not a multiple of 100, or if it is a multiple of 400.

Input data format: A natural number is fed to the program.

Output data format: The program should output the text in accordance with the condition of the problem.