Create Chess_1
You are given two different squares of the chessboard. Write a program that determines whether the rook can get from the first square to the second in one move. The program receives as input four numbers from 1 to 8 each, specifying the column number and line number, first for the first cell, then for the second cell. The program should print "YES" if it is possible to get to the second one from the first square by the move of the rook, or "NO" otherwise.

The format of the input data: The input to the program is four numbers from 1 to 8.

Output data format: The program should output the text in accordance with the condition of the problem.