Create days_quantity
The ordinal number of the month is given (1, \, 2, \ ldots, 12) (1,2, ..., 12). Write a program that displays the number of days in this month. Accept that the year is not a leap year.

Input data format. The program receives one integer at the entrance - the ordinal number of the month.

Output data format. The program should display the number of days in this month.