Create calculator
Write a program that reads two integers and a string from the keyboard. If this line is a designation of one of the four mathematical operations (+, -, *, /), then output the result of applying this operation to the numbers entered earlier, otherwise output "Invalid operation". If the user wants to divide by zero, output the text "You cannot divide by zero!"

Input data format. The input to the program is two integers, each on a separate line, and a string.

Output data format. The program should display the result of applying the operation to the entered numbers or the corresponding text if the operation is incorrect or if division by zero occurs.