Create roulette
On the roulette wheel, the pockets are numbered from 0 to 36. Below are the colors of the pockets:

- pocket 0 green;
- for pockets 1 to 10, the odd numbered pockets are red, the even numbered pockets are black;
- - for pockets from 11 to 18, odd numbered pockets are black, even numbered pockets are red;
for pockets 19 to 28, odd numbered pockets are red, even numbered pockets black;
- for pockets 29 to 36, the odd numbered pockets are black, the even numbered pockets are red.

Write a program that reads the pocket number and indicates whether the pocket is green, red, or black. The program should display an error message if the user enters a number that is outside the range from 0 to 36.

Input data format
The input to the program is one integer.

Output data format
The program should display the color of the pocket or the message "input error" if the entered number is outside the range from 0 to 36.