Create color_mixer
Red, blue and yellow are called primary colors because they cannot be obtained by mixing other colors. When two primary colors are mixed, a secondary color is obtained:

- if you mix red and blue, you get purple;
- if you mix red and yellow, you get orange;
- if you mix blue and yellow, you get green.

Write a program that reads the names of the two primary colors for mixing. If the user enters anything other than red, blue, or yellow, the program should display an error message. Otherwise, the program should print the name of the secondary color, which will be the result.

The format of the input data: Two lines are fed to the program, each on a separate line.

Output format: The program should output the resulting color of the blend or the message "color error" if a color was not entered.